+++ context/app/static/js/components/entity-search/Search/style.js
@@ -1,12 +1,19 @@
 import styled from 'styled-components';
 
-const SearchLayout = styled.div`
+const Flex = styled.div`
   display: flex;
+  margin-bottom: ${(props) => props.theme.spacing(1)}px;
+`;
+
+const Grow = styled.div`
+  flex-grow: 1;
 `;
 
 const ResultsLayout = styled.div`
   flex-grow: 1;
   min-width: 0; // needed for horizontal scrolling table
+  display: flex;
+  flex-direction: column;
 `;
 
-export { SearchLayout, ResultsLayout };
+export { Flex, Grow, ResultsLayout };