+++ context/app/static/js/components/entity-search/SearchWrapper/store.js
@@ -1,20 +1,46 @@
 /* eslint-disable no-param-reassign */
 import create from 'zustand';
-
 import createContext from 'zustand/context';
+
+import immer from 'js/stores/immerMiddleware';
 
 const { Provider, useStore } = createContext();
 
-const createStore = ({ initialFields, initialFacets, fields, facets, filters, entityType }) =>
-  create((set) => ({
-    initialFields,
-    initialFacets,
-    fields,
-    facets,
-    filters,
-    entityType,
-    setFields: (selectedFields) => set({ fields: selectedFields }),
-    setFacets: (selectedFacets) => set({ facets: selectedFacets }),
-  }));
+const createStore = ({
+  initialFields,
+  initialFacets,
+  fields,
+  facets,
+  defaultFilters,
+  entityType,
+  numericFacetsProps,
+}) =>
+  create(
+    immer((set) => ({
+      initialFields,
+      initialFacets,
+      defaultFilters,
+      fields,
+      facets,
+      entityType,
+      numericFacetsProps,
+      setFields: (selectedFields) =>
+        set((state) => {
+          state.fields = selectedFields;
+        }),
+      setFacets: (selectedFacets) =>
+        set((state) => {
+          state.facets = selectedFacets;
+        }),
+      setNumericFacetsProps: (numericFacets) =>
+        set((state) => {
+          state.numericFacetsProps = numericFacets;
+        }),
+      setFacetSize: ({ identifier, size }) =>
+        set((state) => {
+          state.facets[identifier].size = size;
+        }),
+    })),
+  );
 
 export { createStore, Provider, useStore };