+++ context/app/static/js/components/entity-search/facets/facetChips/RangeFacetChip/index.js
@@ -1,0 +1,3 @@
+import RangeFacetChip from './RangeFacetChip';
+
+export default RangeFacetChip;