+++ context/app/static/js/components/entity-search/facets/select/SelectFacet/style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import FormGroup from '@material-ui/core/FormGroup';
+
+const StyledFormGroup = styled(FormGroup)`
+  max-height: 250px;
+  overflow-y: auto;
+  display: block;
+  padding-right: ${(props) => props.theme.spacing(1.5)}px; // Adds space for the scrollbar.
+`;
+
+export { StyledFormGroup };