+++ context/app/static/js/components/entity-search/sidebar/DataTypesSelect/DataTypesSelect.jsx
@@ -3,15 +3,31 @@ import FormGroup from '@material-ui/core/FormGroup';
 import FormControl from '@material-ui/core/FormControl';
 
 import DataTypesSelectCheckboxes from 'js/components/entity-search/sidebar/DataTypesSelectCheckboxes';
-import { StyledPaper } from './style';
+import { StyledPaper, StyledFormLabel } from './style';
 
-function DataTypesSelect({ dataTypesToFieldsMap, handleToggleDataType, groups, handleToggleGroup }) {
+function DataTypesSelect({
+  groups,
+  dataTypesToFieldsMap,
+  handleToggleDataType,
+  handleToggleGroup,
+  selectedGroups,
+  selectedDataTypes,
+}) {
   return (
     <StyledPaper>
       <FormControl component="fieldset">
+        <StyledFormLabel>Data Types</StyledFormLabel>
         <FormGroup aria-label="data-types">
-          <DataTypesSelectCheckboxes values={Object.keys(groups)} eventHandler={handleToggleGroup} />
-          <DataTypesSelectCheckboxes values={Object.keys(dataTypesToFieldsMap)} eventHandler={handleToggleDataType} />
+          <DataTypesSelectCheckboxes
+            values={Object.keys(groups)}
+            eventHandler={handleToggleGroup}
+            selectedValues={selectedGroups}
+          />
+          <DataTypesSelectCheckboxes
+            values={Object.keys(dataTypesToFieldsMap)}
+            eventHandler={handleToggleDataType}
+            selectedValues={selectedDataTypes}
+          />
         </FormGroup>
       </FormControl>
     </StyledPaper>