+++ context/app/static/js/components/entity-search/sidebar/Sidebar/style.js
@@ -1,7 +1,8 @@
 import styled from 'styled-components';
 
 const SidebarLayout = styled.div`
-  width: 20%;
+  width: 250px;
+  min-width: 250px;
   margin-right: ${(props) => props.theme.spacing(2)}px;
 `;