+++ context/app/static/js/helpers/trackers.js
@@ -5,7 +5,7 @@ function getSiteId(location) {
   const { host } = location;
 
   // siteIds should correspond to the list at:
-  // https://hubmapconsortium.matomo.cloud/index.php?module=SitesManager
+  // https://hubmap.matomo.cloud/index.php?module=SitesManager
 
   switch (host) {
     case 'portal.hubmapconsortium.org':
@@ -18,13 +18,13 @@ function getSiteId(location) {
 }
 
 const tracker = new MatomoTracker({
-  urlBase: 'https://hubmapconsortium.matomo.cloud/',
+  urlBase: 'https://hubmap.matomo.cloud/',
   // eslint-disable-next-line no-restricted-globals
   siteId: getSiteId(location),
   // userId: 'UID76903202', // optional, default value: `undefined`.
-  // trackerUrl: 'https://LINK.TO.DOMAIN/tracking.php', // optional, default value: `${urlBase}matomo.php`
-  // srcUrl: 'https://LINK.TO.DOMAIN/tracking.js', // optional, default value: `${urlBase}matomo.js`
-  // disabled: false, // optional, false by default. Makes all tracking calls no-ops if set to true.
+  // trackerUrl: 'https://hubmap.matomo.cloud/tracking.php', // optional, default value: `${urlBase}matomo.php`
+  // srcUrl: 'https://hubmap.matomo.cloud/tracking.js', // optional, default value: `${urlBase}matomo.js`
+  disabled: process.env.NODE_ENV === 'test', // Tracking calls should be no-ops during tests.
   // heartBeat: { // optional, enabled by default
   //   active: true, // optional, default value: true
   //   seconds: 10 // optional, default value: `15
@@ -38,7 +38,7 @@ const tracker = new MatomoTracker({
   // }
 });
 
-ReactGA.initialize('UA-133341631-3');
+ReactGA.initialize('UA-133341631-3', { testMode: process.env.NODE_ENV === 'test' });
 
 function trackPageView(path) {
   tracker.trackPageView({ href: path });
@@ -66,11 +66,16 @@ function trackSiteSearch(keyword) {
   tracker.trackSiteSearch({
     keyword,
   });
+  /*
+  We currently call both trackEvent and trackSiteSearch:
+  while that is the case, we don't have to double track the even in GA.
+  
   ReactGA.trackEvent({
     // category: analyticsCategory,
     action: 'Free Text Search',
     label: keyword,
   });
+  */
 }
 
 export { trackPageView, trackEvent, trackLink, trackSiteSearch };