+++ context/app/static/js/shared-styles/inputs/SearchBar/style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+import TextField from '@material-ui/core/TextField';
+
+const StyledTextField = styled(TextField)`
+  & > :first-child {
+    border-radius: 4px;
+  }
+`;
+
+export { StyledTextField };